Moved Class VTOLMOD
