+++ ModLoader/ModLoader.csproj
@@ -94,6 +94,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SPModManager.cs" />
     <Compile Include="UConsole.cs" />
+    <Compile Include="VTOLMOD.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
